Update README.md
This Python program tracks price of a product on Amazon website(User defined). Prices of the products on Amazon changes every minute according to the demand and supply of the products. Whenever the price of the product fall below a desired level(set by the user).  It will send an email alert to the user by logging into a account. through googles less secure app access and send alerts to one or multiple email id's as desired.